+++ sys/arch/i386/include/_types.h
@@ -1,4 +1,4 @@
-/*	$OpenBSD: _types.h,v 1.12 2011/09/08 02:47:13 guenther Exp $	*/
+/*	$OpenBSD: _types.h,v 1.13 2011/09/08 03:40:32 guenther Exp $	*/
 
 /*-
  * Copyright (c) 1990, 1993
@@ -51,7 +51,7 @@
 
 #if defined(_KERNEL)
 typedef struct label_t {
-	int val[6];
+	long val[6];
 } label_t;
 #endif
